Rename orphancompilers.py to propscheck.py
The naming fits better
