+++ .husky/pre-commit
@@ -4,4 +4,4 @@
 make prereqs
 npx lint-staged
 npm run test
-python3 ./etc/scripts/util/orphancompiler.py
+python3 ./etc/scripts/util/propscheck.py
