+++ etc/scripts/util/propscheck.py
