+++ lib/compiler-finder.js
@@ -280,7 +280,7 @@ export class CompilerFinder {
             return [];
         }
 
-        // This seems like no longer needed with orphancompiler.py checks
+        // This seems like no longer needed with propscheck.py checks
         compilerInfo.libPath = compilerInfo.libPath.split(path.delimiter).filter(p => p !== '');
 
         logger.debug('Found compiler', compilerInfo);
